refactor sending email to smaller services
